+++ applications/fishing-map/src/features/workspace/context-areas/ContextAreaLayerPanel.tsx
@@ -57,9 +57,7 @@ function LayerPanel({ dataview, onToggle = () => {} }: LayerPanelProps): React.R
     return <DatasetNotFound dataview={dataview} />
   }
 
-  const title = isUserLayer
-    ? dataset?.name || dataset?.id
-    : t(`datasets:${dataset?.id}.name` as any)
+  const title = t(`datasets:${dataset?.id}.name` as any, dataset?.name || dataset?.id)
 
   const TitleComponent = (
     <Title