+++ packages/timebar/types/timebar.d.ts
@@ -10,6 +10,7 @@ declare class Timebar extends Component<any, any, any> {
     getMinimumRangeMs: (this: any, minimumRange: any, minimumRangeUnit: any) => any;
     toggleTimeRangeSelector: () => void;
     setBookmark: () => void;
+    setLocale: (this: any, locale: any) => any;
     onTimeRangeSelectorSubmit: (start: any, end: any) => void;
     zoom: (zoom: any) => void;
     notifyChange: (start: any, end: any, source: any, clampToEnd?: boolean) => void;
@@ -41,6 +42,7 @@ declare namespace Timebar {
         const maximumRange: any;
         const maximumRangeUnit: any;
         const showLastUpdate: any;
+        const locale: any;
     }
     namespace defaultProps {
         const latestAvailableDataDate_1: string;
@@ -112,6 +114,8 @@ declare namespace Timebar {
         export { maximumRangeUnit_1 as maximumRangeUnit };
         const showLastUpdate_1: boolean;
         export { showLastUpdate_1 as showLastUpdate };
+        const locale_1: string;
+        export { locale_1 as locale };
     }
 }
 import { Component } from "react";