+++ applications/dataviews-editor/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @globalfishingwatchapp/dataviews-editor
 
+## 1.1.75
+
+### Patch Changes
+
+- Updated dependencies [58d28355]
+  - @globalfishingwatch/dataviews-client@11.8.0
+  - @globalfishingwatch/layer-composer@8.10.0
+
 ## 1.1.74
 
 ### Patch Changes