+++ packages/layer-composer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @globalfishingwatch/layer-composer
 
+## 8.10.0
+
+### Minor Changes
+
+- 58d28355: Handle user context layers with line features
+
 ## 8.9.1
 
 ### Patch Changes